Add files via upload
This directory contains the code and some examples for doing the Boozer transform of VMEC data and creating the input file (woutb) that FAR3d requires for its equilibrium data. The source code directory should replace the usual one provided in STELLOPT and is compiled using the STELLOPT setup script.