Add files via upload
These changes were made by Luis Garcia (of the Universidad Carlos III, Madrid, Spain) to correct errors in how the ion/electron damping and thermal ion FLR terms were introduced in FAR3d. These errors in the earlier code led not only to incorrect damping rates, but also in some cases resulted in SegFaults. The following is a summary of the changes that were made:
- The Landau ion/electron damping terms were modified to correct errors in the earlier version
- This involved adding two new subroutines: blockjl and b2lxl
- Corrections were made in the thermal ion FLR terms, specifically in calls to dlsq_r and related subroutines
- To calculate dlsq_r and the corresponding blocks, two equilibrium matrices: lplr_r and lplt_r were introduced
- The code has been further simplified by removing some redundant equilibrium matrices
- Note that the Trapped_on and vtheq terms are not included
- Also, the inputlist_namelist.f90 file was modified to be consistent with the new version