Add files via upload
These updates to the FAR3d code add the capability of specifying the ion mass for the two energetic particle components. spe1 is the ion to proton mass for the first EP species (usually a beam) and spe2 is the ion to proton mass for the 2nd EP species (usually alpha particles).